+++ src/components/banner/GovBanner/GovBanner.tsx
@@ -7,7 +7,7 @@ import { BannerContent } from '../BannerContent/BannerContent'
 import { BannerGuidance } from '../BannerGuidance/BannerGuidance'
 import { BannerIcon } from '../BannerIcon/BannerIcon'
 import { MediaBlockBody } from '../../mediablock/MediaBlockBody/MediaBlockBody'
-import { BannerLockImage } from '../BannerLockImage/BannerLockImage'
+import { Icon } from '../../Icon/Icons'
 
 // assets
 import flagImg from '@uswds/uswds/src/img/us_flag_small.png'
@@ -20,6 +20,7 @@ type TLD = '.gov' | '.mil'
 
 interface GovBannerCopy {
   header: string
+  ariaLabel: string
   headerAction: string
   tldSectionHeader: string
   tldSectionContent: JSX.Element
@@ -28,12 +29,11 @@ interface GovBannerCopy {
 }
 
 const getCopy = (language: Language, tld: TLD): GovBannerCopy => {
-  const lock = <BannerLockImage title="Lock" description="A locked padlock" />
-
   switch (language) {
     case 'english':
       return {
         header: 'An official website of the United States government',
+        ariaLabel: 'Official website of the United States government',
         headerAction: 'Here’s how you know',
         tldSectionHeader: `Official websites use ${tld}`,
         tldSectionContent: ((): JSX.Element => {
@@ -57,15 +57,20 @@ const getCopy = (language: Language, tld: TLD): GovBannerCopy => {
         httpsSectionHeader: `Secure ${tld} websites use HTTPS`,
         httpsSectionContent: (
           <>
-            A <strong>lock ( {lock} )</strong> or <strong>https://</strong>{' '}
-            means you’ve safely connected to the {tld} website. Share sensitive
-            information only on official, secure websites.
+            A{' '}
+            <strong>
+              lock (<Icon.Lock aria-label="Locked padlock icon" />)
+            </strong>{' '}
+            or <strong>https://</strong> means you’ve safely connected to the{' '}
+            {tld} website. Share sensitive information only on official, secure
+            websites.
           </>
         ),
       }
     case 'spanish':
       return {
         header: 'Un sitio oficial del Gobierno de Estados Unidos',
+        ariaLabel: 'Un sitio oficial del Gobierno de Estados Unidos',
         headerAction: 'Así es como usted puede verificarlo',
         tldSectionHeader: `Los sitios web oficiales usan ${tld}`,
         tldSectionContent: ((): JSX.Element => {
@@ -89,10 +94,13 @@ const getCopy = (language: Language, tld: TLD): GovBannerCopy => {
         httpsSectionHeader: `Los sitios web seguros ${tld} usan HTTPS`,
         httpsSectionContent: (
           <>
-            Un <strong>candado ( {lock} )</strong> o <strong>https://</strong>{' '}
-            significa que usted se conectó de forma segura a un sitio web {tld}.
-            Comparta información sensible sólo en sitios web oficiales y
-            seguros.
+            Un{' '}
+            <strong>
+              candado (<Icon.Lock aria-label="Icono de candado cerrado" />)
+            </strong>{' '}
+            o <strong>https://</strong> significa que usted se conectó de forma
+            segura a un sitio web {tld}. Comparta información sensible sólo en
+            sitios web oficiales y seguros.
           </>
         ),
       }
@@ -112,22 +120,34 @@ export const GovBanner = ({
 }: GovBannerProps & JSX.IntrinsicElements['section']): ReactElement => {
   const [isOpen, setIsOpen] = useState(false)
 
-  const copy = getCopy(language, tld)
+  const {
+    header,
+    ariaLabel,
+    headerAction,
+    httpsSectionHeader,
+    httpsSectionContent,
+    tldSectionHeader,
+    tldSectionContent,
+  } = getCopy(language, tld)
 
   return (
-    <Banner className={className} data-testid="govBanner" {...sectionProps}>
+    <Banner
+      className={className}
+      data-testid="govBanner"
+      aria-label={ariaLabel}
+      {...sectionProps}>
       <BannerHeader
         isOpen={isOpen}
-        flagImg={<BannerFlag src={flagImg} alt="U.S. flag" />}
-        headerText={copy.header}
-        headerActionText={copy.headerAction}>
+        flagImg={<BannerFlag src={flagImg} aria-hidden alt="" />}
+        headerText={header}
+        headerActionText={headerAction}>
         <BannerButton
           isOpen={isOpen}
           aria-controls="gov-banner"
           onClick={(): void => {
             setIsOpen((previousIsOpen) => !previousIsOpen)
           }}>
-          {copy.headerAction}
+          {headerAction}
         </BannerButton>
       </BannerHeader>
       <BannerContent id="gov-banner" isOpen={isOpen}>
@@ -136,9 +156,9 @@ export const GovBanner = ({
             <BannerIcon src={dotGovIcon} alt="" />
             <MediaBlockBody>
               <p>
-                <strong>{copy.tldSectionHeader}</strong>
+                <strong>{tldSectionHeader}</strong>
                 <br />
-                {copy.tldSectionContent}
+                {tldSectionContent}
               </p>
             </MediaBlockBody>
           </BannerGuidance>
@@ -146,9 +166,9 @@ export const GovBanner = ({
             <BannerIcon src={httpsIcon} alt="" />
             <MediaBlockBody>
               <p>
-                <strong>{copy.httpsSectionHeader}</strong>
+                <strong>{httpsSectionHeader}</strong>
                 <br />
-                {copy.httpsSectionContent}
+                {httpsSectionContent}
               </p>
             </MediaBlockBody>
           </BannerGuidance>