+++ src/components/forms/CharacterCount/CharacterCount.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useRef } from 'react'
 import classnames from 'classnames'
 
 import { TextInput, TextInputProps } from '../TextInput/TextInput'
@@ -37,6 +37,7 @@ type BaseCharacterCountProps = {
   id: string
   name: string
   maxLength: number
+  value?: string
   defaultValue?: string
   className?: string
   isTextArea?: boolean
@@ -57,6 +58,7 @@ export const CharacterCount = ({
   name,
   className,
   maxLength,
+  value = '',
   defaultValue = '',
   isTextArea = false,
   getCharacterCount = defaultCharacterCount,
@@ -65,21 +67,27 @@ export const CharacterCount = ({
 }:
   | TextInputCharacterCountProps
   | TextareaCharacterCountProps): React.ReactElement => {
-  const initialCount = getCharacterCount(defaultValue)
+  const initialCount = getCharacterCount(value || defaultValue)
   const [length, setLength] = useState(initialCount)
   const [message, setMessage] = useState(getMessage(initialCount, maxLength))
   const [isValid, setIsValid] = useState(initialCount < maxLength)
+  const srMessageRef = useRef<HTMLDivElement>(null)
 
   const classes = classnames('usa-character-count__field', className)
-  const messageClasses = classnames(
-    'usa-hint',
-    'usa-character-count__message',
-    { 'usa-character-count__message--invalid': !isValid }
-  )
+  const messageClasses = classnames('usa-hint', 'usa-character-count__status', {
+    'usa-character-count__status--invalid': !isValid,
+  })
 
   useEffect(() => {
-    setMessage(getMessage(length, maxLength))
+    const message = getMessage(length, maxLength)
+    setMessage(message)
     setIsValid(length <= maxLength)
+    // Updates the character count status for screen readers after a 1000ms delay
+    const timer = setTimeout(() => {
+      // Setting the text directly for VoiceOver compatibility.
+      if (srMessageRef.current) srMessageRef.current.textContent = message
+    }, 1000)
+    return () => clearTimeout(timer)
   }, [length])
 
   const handleBlur = (
@@ -114,18 +122,19 @@ export const CharacterCount = ({
     const { onBlur, onChange, inputRef, ...textAreaProps } =
       remainingProps as Partial<TextareaCharacterCountProps>
 
-    InputComponent = (
-      <Textarea
-        id={id}
-        name={name}
-        className={classes}
-        defaultValue={defaultValue}
-        onBlur={(e): void => handleBlur(e, onBlur)}
-        onChange={(e): void => handleChange(e, onChange)}
-        inputRef={inputRef}
-        {...textAreaProps}
-      />
-    )
+    const attributes = {
+      id: id,
+      name: name,
+      className: classes,
+      ...(value ? { value: value } : { defaultValue: defaultValue }),
+      onBlur: (e: React.FocusEvent<HTMLTextAreaElement, Element>): void =>
+        handleBlur(e, onBlur),
+      onChange: (e: React.ChangeEvent<HTMLTextAreaElement>): void =>
+        handleChange(e, onChange),
+      inputRef: inputRef,
+      ...textAreaProps,
+    }
+    InputComponent = <Textarea {...attributes} />
   } else {
     const {
       onBlur,
@@ -134,32 +143,40 @@ export const CharacterCount = ({
       type = 'text',
       ...inputProps
     } = remainingProps as Partial<TextInputCharacterCountProps>
-
-    InputComponent = (
-      <TextInput
-        id={id}
-        type={type}
-        name={name}
-        className={classes}
-        defaultValue={defaultValue}
-        onBlur={(e): void => handleBlur(e, onBlur)}
-        onChange={(e): void => handleChange(e, onChange)}
-        inputRef={inputRef}
-        {...inputProps}
-      />
-    )
+    const attributes = {
+      id: id,
+      type: type,
+      name: name,
+      className: classes,
+      ...(value ? { value: value } : { defaultValue: defaultValue }),
+      onBlur: (e: React.FocusEvent<HTMLInputElement, Element>): void =>
+        handleBlur(e, onBlur),
+      onChange: (e: React.ChangeEvent<HTMLInputElement>): void =>
+        handleChange(e, onChange),
+      inputRef: inputRef,
+      ...inputProps,
+    }
+    InputComponent = <TextInput {...attributes} />
   }
 
   return (
     <>
       {InputComponent}
-      <span
-        data-testid="characterCountMessage"
-        id={`${id}-info`}
-        className={messageClasses}
-        aria-live="polite">
-        {message}
+      <span className="usa-sr-only" id={`${id}-info`}>
+        You can enter up to {maxLength} characters
       </span>
+      <div
+        className={messageClasses}
+        aria-hidden="true"
+        data-testid="characterCountMessage">
+        {message}
+      </div>
+      <div
+        ref={srMessageRef}
+        className="usa-character-count__sr-status usa-sr-only"
+        aria-live="polite"
+        data-testid="characterCountSRMessage"
+      />
     </>
   )
 }